Move softfloat into dependencies directory
Move the softfloat library to a dependencies directory. In future we can add GMP there too to reduce the number of external dependencies needed and to support static linking.

The directory was renamed from `SoftFloat-3e` to `berkeley-softfloat-3` because that's how the latest version is named.

It hasn't been upgraded in this commit to make it simpler to review and help Git track history. A future commit will upgrade it.