Add a DOWNLOAD_GMP option to build against a local GMP build
This downloads the latest release of GMP and builds it as part of the overall emulator build. This should make it easier to build sail without having to install system dependencies.

Also change `build_simulators.sh` to download GMP by default to make it easier for users that don't manually configure the CMake build.

Finally change `build_simulators.sh` to detect the number of CPUs to use since it speeds up compiling GMP.